Fold the ThreadDiagnostics facade into the factory translation unit
The forwarder was a 10-line file of its own. It has no state and no other
callers, so it lives next to PeerConnectionDependencyFactory::InFlightDispatches()
which it wraps. Drops one source file and one BUILD.gn entry.

Also names pc_thread in the InFlightDispatch doc comments, which still described
the version that reported only libwebrtc's three internal threads.
